requirements.txt
pickle5
python-chess
pillow